--- .github/workflows/secretscan.yml
@@ -1,27 +1,0 @@
-# This is a basic workflow to help you get started with Actions
-
-name: Secret Detector
-
-# Controls when the action will run. Triggers the workflow on push or pull request
-# events but only for the master branch
-on:
-  push:
-    branches:
-      - '**'
-
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
-jobs:
-  # This workflow contains a single job called "build"
-  build:
-    # The type of runner that the job will run on
-    runs-on: ubuntu-latest
-
-    # Steps represent a sequence of tasks that will be executed as part of the job
-    steps:
-    # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-    - uses: actions/checkout@v2
-
-    # Scan for secrets in the repo
-    - name: Secret Detector
-      uses: evanextreme/detect-secrets-action@1.0.0
-
