+++ functions/helpers/checkAuth.js
