+++ functions/post/index.js
@@ -1,6 +1,6 @@
 const express = require("express")
-const checkAuth = require("../checkAuth")
-const checkRecaptcha = require("../checkRecaptcha")
+const checkAuth = require("../helpers/checkAuth")
+const checkRecaptcha = require("../helpers/checkRecaptcha")
 const router = express.Router()
 const createPost = require("./createPost")
 const deletePost = require("./deletePost")
